Refacto CMEditViewLocalePicker
