+++ packages/strapi-plugin-i18n/admin/src/selectors/selectI18nLocales.js
@@ -1,0 +1,3 @@
+const selectI18NLocales = state => state.get('i18n_locales').locales;
+
+export default selectI18NLocales;
